Update index.html
The two changes are to really clearly refer to the Start Button light and Reset Button Light in Table 2 to clearly differentiate between the Start and Reset state change commands. The other items is the Held Immenent Warning is unclear I would simply make it Hold Imminent or Hold Warning.